+++ plugin/storage/influxdb/factory.go
@@ -23,6 +23,8 @@ import (
 	"github.com/spf13/viper"
 	"github.com/uber/jaeger-lib/metrics"
 	"go.uber.org/zap"
+	"net/http"
+	"time"
 )
 
 type Factory struct {
@@ -36,23 +38,28 @@ type Factory struct {
 func (f *Factory) Initialize(metricsFactory metrics.Factory, logger *zap.Logger) error {
 	f.metricsFactory, f.logger = metricsFactory, logger
 	logger.Info("Factory Initialize", zap.String("type", "influxdb"))
-
 	if f.options.TLS.Enabled {
-		tls, err := f.options.TLS.Config(logger)
+		tls, err := f.options.TLS.Config(f.logger)
 		if err != nil {
-			logger.Panic("please check TLS Setting", zap.Error(err))
+			return err
 		}
 		f.options.Options.SetTLSConfig(tls)
 	}
+	f.options.SetHTTPClient(
+		&http.Client{
+			Transport: newLatencyTransport(http.DefaultTransport, metricsFactory),
+			Timeout:   time.Second * time.Duration(f.options.HTTPRequestTimeout()),
+		})
+	err := f.options.checkNecessaryOptions()
+	if err != nil {
+		return err
+	}
 	f.client = influxdb2.NewClientWithOptions(f.options.url, f.options.token, &f.options.Options)
-
-	//client.
 	return nil
 }
 
 func (f *Factory) CreateWriter() (metricstore.Writer, error) {
-	//TODO implement me
-	panic("implement me")
+	return NewMetricWriter(f.logger, f.client, f.options.org, f.options.bucket), nil
 }
 
 func NewFactory() *Factory {
@@ -74,4 +81,30 @@ func (f *Factory) InitFromViper(v *viper.Viper) {
 // InitFromOptions initializes factory from options.
 func (f *Factory) InitFromOptions(o Options) {
 	f.options = o
+}
+
+type latencyTransport struct {
+	transport http.RoundTripper
+	latency   metrics.Timer
+	errors    metrics.Counter
+}
+
+func (l *latencyTransport) RoundTrip(request *http.Request) (*http.Response, error) {
+	now := time.Now()
+	resp, err := l.transport.RoundTrip(request)
+	if err != nil {
+		l.errors.Inc(1)
+		return resp, err
+	}
+	l.latency.Record(time.Since(now))
+	return resp, err
+}
+
+func newLatencyTransport(t http.RoundTripper, f metrics.Factory) http.RoundTripper {
+	m := f.Namespace(metrics.NSOptions{Name: "influxdb", Tags: nil})
+	return &latencyTransport{
+		transport: t,
+		latency:   m.Timer(metrics.TimerOptions{Name: "latency", Tags: nil}),
+		errors:    m.Counter(metrics.Options{Name: "errors", Tags: nil}),
+	}
 }